refactor(workstation): route Code Editor tabs through UnifiedTabContent
Pre-commit hook ran. Total eslint: 0, total circular: 0
