+++ src/modules/WorkStation/TabContent/renderers/lintScan.tsx
@@ -1,22 +1,22 @@
 /**
  * Renderer wrapper for `lint-scan` tabs.
  *
- * `LintScanContent` needs a `repoPath` prop. We resolve it from
- * the active workspace root so the wrapper can run outside the editor host —
- * which matches how the live editor renderer derives it today. The
- * dispatcher itself is not yet wired in (Phase 2), so this is a
- * preparatory adapter.
+ * `LintScanContent` needs a `repoPath` prop. We pull it from the hoisted Code
+ * Editor host context (`useEditorHostContext`) for strict parity with
+ * `TabContentRenderer`'s `case "lint-scan"`, which passed the host's selected
+ * `repoPath` (not the active workspace root). `lint-scan` is editor-only, so
+ * coupling this renderer to the editor host is safe — it throws if mounted
+ * outside an `EditorHostProvider`.
  */
-import { useAtomValue } from "jotai";
 import React, { memo } from "react";
 
 import LintScanContent from "@src/modules/WorkStation/CodeEditor/Panels/EditorMainPane/content/LintScanContent";
-import { activeWorkspaceRootPathAtom } from "@src/store/workspace";
+import { useEditorHostContext } from "@src/modules/WorkStation/CodeEditor/Panels/EditorMainPane/context/editorHostContext";
 
 import type { UnifiedTabContentProps } from "../types";
 
 const LintScanTabRenderer: React.FC<UnifiedTabContentProps> = memo(() => {
-  const repoPath = useAtomValue(activeWorkspaceRootPathAtom);
+  const { repoPath } = useEditorHostContext();
   return <LintScanContent repoPath={repoPath} />;
 });
